+++ common/src/main/java/de/maxhenkel/voicechat/plugins/PluginManager.java
@@ -17,6 +17,7 @@ import de.maxhenkel.voicechat.voice.common.*;
 import de.maxhenkel.voicechat.voice.server.Group;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.phys.Vec3;
 
 import javax.annotation.Nullable;
 import java.util.List;
@@ -59,7 +60,7 @@ public class PluginManager {
         events = eventBuilder.build();
     }
 
-    public <T extends Event> boolean dispatchEvent(Class<T> eventClass, T event) {
+    public <T extends Event> boolean dispatchEvent(Class<? extends T> eventClass, T event) {
         List<Consumer<? extends Event>> events = this.events.get(eventClass);
         if (events == null) {
             return false;
@@ -186,6 +187,31 @@ public class PluginManager {
             return null;
         }
         return clientSoundEvent.getRawAudio();
+    }
+
+    public void onALSound(int source, @Nullable UUID channelId, @Nullable Vec3 pos, Class<? extends OpenALSoundEvent> eventClass) {
+        dispatchEvent(eventClass, new OpenALSoundEventImpl(
+                new VoicechatClientApiImpl(),
+                channelId,
+                pos == null ? null : new PositionImpl(pos),
+                source
+        ));
+    }
+
+    public void onCreateALContext(long context, long device) {
+        dispatchEvent(CreateOpenALContextEvent.class, new CreateOpenALContextEventImpl(
+                new VoicechatClientApiImpl(),
+                context,
+                device
+        ));
+    }
+
+    public void onDestroyALContext(long context, long device) {
+        dispatchEvent(DestroyOpenALContextEvent.class, new DestroyOpenALContextEventImpl(
+                new VoicechatClientApiImpl(),
+                context,
+                device
+        ));
     }
 
     private static PluginManager instance;